+++ src/dictionary/routes.js
@@ -3,7 +3,9 @@ import w from './words';
 
 const routes = {
     login: `/${w.login}`,
-    mainMenu: '/'
+    companyList: '/',
+    companyRoot: '/companies/:companyId',
+    campaignRoot: '/campaigns/:campaignId'
 }
 
 export default routes;